Aktualizácia projektu s novými funkciami a opravami

+++ tasks/views.py
@@ -1,17 +1,15 @@
-# tasks/views.py
-
-from django.shortcuts import render, redirect
-from django.contrib.auth.decorators import login_required
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Task
 from .forms import TaskForm
+from django.contrib.auth.decorators import login_required
 
 @login_required
 def task_list(request):
-    tasks = Task.objects.filter(user=request.user).order_by('start_time')
+    tasks = Task.objects.filter(user=request.user)
     return render(request, 'tasks/task_list.html', {'tasks': tasks})
 
 @login_required
-def create_task(request):
+def task_create(request):
     if request.method == 'POST':
         form = TaskForm(request.POST)
         if form.is_valid():
@@ -24,24 +22,27 @@ def create_task(request):
     return render(request, 'tasks/task_form.html', {'form': form})
 
 @login_required
-def update_task(request, pk):
-    task = Task.objects.get(pk=pk)
+def task_complete(request, task_id):
+    task = get_object_or_404(Task, id=task_id, user=request.user)
+    if request.method == 'POST':
+        task.delete()  # Zmena pre vymazanie úlohy
+    return redirect('task_list')
+
+@login_required
+def task_edit(request, task_id):
+    task = get_object_or_404(Task, id=task_id, user=request.user)
     if request.method == 'POST':
         form = TaskForm(request.POST, instance=task)
         if form.is_valid():
-            form.save()
-            return redirect('task_list')
+            if 'completed' in request.POST and request.POST['completed'] == 'on':
+                task.delete()
+                return redirect('task_list')
+            else:
+                form.save()
+                return redirect('task_list')
     else:
         form = TaskForm(instance=task)
-    return render(request, 'tasks/task_form.html', {'form': form})
-
-@login_required
-def delete_task(request, pk):
-    task = Task.objects.get(pk=pk)
-    if request.method == 'POST':
-        task.delete()
-        return redirect('task_list')
-    return render(request, 'tasks/task_confirm_delete.html', {'task': task})
+    return render(request, 'tasks/task_form.html', {'form': form, 'task': task})
 
 def index(request):
     return render(request, 'index.html')
